pyproject.toml
[project]
name = "japanese-font-library-presetbuilder"
version = "0.1.0"
description = "Project for developing Japanese fonts presetbuilder for Skyrim font mods."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyside6>=6.10.2",
    "pyyaml>=6.0.3",
]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
